Detect cyclic objects when marshaling objects.
Marionette does currently not test for cyclic object references as
it marshals return values for transport across the wire.

Example of cyclic object:

	let obj = {};
	obj.cyclic = obj;

Passing this through evalaute.toJSON currently causes an infinite
recursion due to obj being referenced inside itself.  We can use
JSON.stringify to test if obj contains such cyclic values.  It is
assumed that the input to assert.acyclic is already JSON safe, so it can
be parsed by JSON.stringify, because of the previous checks it has made.
bugzilla-url: https://bugzilla-dev.allizom.org/show_bug.cgi?id=Bug 1106913
gecko-commit: 04b16dc7a9b8a429049fdfef93436483c2ef662b
gecko-integration-branch: autoland
gecko-reviewers: whimboo