Prevent missing click event from causing deadlock in interaction.flushEventLoop.
If the DOM click event does not fire, which currently happens in
this edge case where Firefox does not bubble the click event up
to <body>, interaction.flushEventLoop will hang forever.

This is believed to be a bug in Gecko, but it is probably safer to
use a TimedPromise here in case anything else bad were to happen.

Thanks-to: Alexei Barantsev <barancev@gmail.com>
bugzilla-url: https://bugzilla-dev.allizom.org/show_bug.cgi?id=Bug 1417821
gecko-commit: 12a1b7b8eb069803f2f9fc80a2d7df0baaae2878
gecko-integration-branch: central
gecko-reviewers: whimboo