Use classic pytest output style for WPT.
The upstream upgrade of pytest came with a new test progress
percentage feature that introduced a lot of unnecessary log output
in WPT wdspec tests.

This adds a pytest flag that reverts it to use the "classic" output
style, which does not have a test progress bar.
bugzilla-url: https://bugzilla-dev.allizom.org/show_bug.cgi?id=1429388
gecko-commit: 4f065bc82a465371aaa8b6dc0de9514cf6e859c8
gecko-integration-branch: central
gecko-reviewers: davehunt, jgraham