Safely shutdown Firefox from in delete_session.
With the request to shutdown the browser, a given amount of time
has to be waited to allow the process to shutdown itself. Only
if the process is still running afterward it has to be killed.

Firefox has an integrated background monitor which observes
long running threads during shutdown, and kills those after
65s. To allow Firefox to shutdown on its own, geckodriver
has to wait that time, and some additional seconds.
bugzilla-url: https://bugzilla-dev.allizom.org/show_bug.cgi?id=1403923
gecko-commit: 025d4690bfbe602a82d91f827a67208df80dc110
gecko-integration-branch: central
gecko-reviewers: jgraham