Introduce assert_time_equals_literal and use it.
This assertion is supposed to be used where the first argument has a tolerance
but the second argument doesn't have such tolerance.  Whereas
assert_times_equal() is supposed to be used for the case both arguments have
the same tolerance, actually it hasn't, it will be fixed in a subsequent patch
in this patch series.
bugzilla-url: https://bugzilla-dev.allizom.org/show_bug.cgi?id=1430654
gecko-commit: edbac2302f907df4fca6aea3843bd033a1a1bb89
gecko-integration-branch: central
gecko-reviewers: birtles